refactor: move energy calculator to application
EnergyCalculator represents an application specific usage of the eCalc
domain.
