+++ src/libecalc/application/energy_calculator.py
