+++ src/tests/libecalc/integration/test_multiple_installations.py
@@ -2,9 +2,9 @@ from datetime import datetime
 
 import pytest
 from libecalc import dto
+from libecalc.application.energy_calculator import EnergyCalculator
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import RateType, TimeSeriesRate
-from libecalc.core.energy_calculator import EnergyCalculator
 from libecalc.core.graph_result import GraphResult
